requirements.txt
fastapi
streamlit
pandas
Pillow